refactor(linter/plugins): move comments methods into own file (#20363)
Pure refactor. Move `sourceCode` methods related to comments into a separate file `comments_methods.ts`. Does not alter any of the code, just moves it.
